2019 Days 09,11,13,15,17,19
These days all use the intcode interpreter. Intcode.py contains a class that acts as a full intcode computer as well as a function for loading intcode files. It also contains several test cases and solutions for days 2, 5, and 9. The class is subclassed for other problems including 11, 13, 15, and 17. (9 and 19 use the unmodified intcode computer)