2018 Day 10
A bit of fluffing about to find where the points converged, then sliding the slider until they formed text. Won't work in the general case unless you change the xy bounds (lines 24,25) and the min max times (lines 32,33). init_t (line 21) is set for my input (and just happens to be the answer to part 2)