2018 Day 19
Once again had to reverse engineer a made up language. Turns out it was a program to find the sum of factors of 10551430. Original code was O(n^2). Reverse engineered code was simplified to O(n) [19c.py].

Files in this commit:
 - 19.py: First attempt. Works for part 1 (~8s) but not part 2
 - 19.txt: Original puzzle input
 - 19-annotated.txt: Puzzle input with notes about what each part does, abstracted to loops
 - 19b.py: Python transpile of 19-annotated.txt: O(n^2)
 - 19c.py: A much simpler and faster approach for both parts (Uncomment line3 for part 2): O(n)
NOTE: Once again, this is specific to my input. The values of r2 may be different for other inputs.