2018 Day 21
Had to reverse engineer the program to solve both parts. 21c.py answers both parts. The approach to part 2 was essentially what is the last number before there is a duplicate. As always, the answers commented in the files are for my input. You may need to rewrite everything for your own input.