2017 Day 25
Note that this solution does not work for all inputs. If you are using this, you will need to change the CHECKSUM_STEPS and the TuringMachine class. The structure of the TuringMachine class is exactly the same as the input structure. Adding the slot to the tape is the equivalent of write 1 and discarding the slot is the equivalent of write 0. Right increments slot and left decrements slot. TuringMachine.state refers to the next state method. I felt that solving it this way would be faster than writing a program to parse the input.