2021 Day 6 UPDATE
Move from using a dict to a list as that will use less memory and prevent the need for an inner loop (just use a couple of list slices and no temp variable). Also remove the initial approach to part 1 since it was kinda slow (and I know this approach works, so I don't need the first one for verification). This may be slightly less readable than the previous version, but it works! (and it's much shorter)