2021 Day 8
Part 1 was fairly fast. Part 2 was tedious, but I am happy that it works. I'm sure deduce_line could be improved, but it works and it works quickly. The general approach was to eliminate impossible candidates until there is only one wire candidate left for each digit. After that, wires are mapped back to display segments, which are then sorted and mapped back to numbers.