2019 Day 22
Note that this was not the code initially used to solve part 1 (For part 1, I actually shuffled list(range(10007)) following the instructions)
I plotted the positions of the cards after a shuffle, which formed straight lines. Using this pattern and a lot of trial and error, I started working on a mathematical solution.
This code builds up a few layers of efficiency, starting by strictly following the shuffle on a single card (I don't have enough RAM for 119315717514047 cards), then finding the inverse of this process. Then the position of any card after the shuffle can be found given the positions of the first two cards.
Cards can be shuffled multiple times with a loop, but it is faster to shuffle them using maths (looping 101741582076661 times is a bit too slow).
The position of card 2020 can be unshuffled using the inverse of fast_multi_shuffle, solving part 2