2021 Day 15
Part 2 is solved by setting the variable PART2 to True on line 26. It is kind of chance that it actually solves it eventually. I'm sure someone could find an input that this doesn't work for, but that could probably be 'solved' by changing the loop on line 40. I assume that there is no need to enter the same space twice as that would always be a worse path.