2021 Day 17
Basically an exhaustive search of possible x,y velocities that end up in the target area. Not the fastest, but it was faster to wait it out than it would be to optimise. The actual used x values go from 11 to 94 (code searches 0 to 94); the y values go from -171 to 170 (code searches -171 to 1000 - not sure the best way to find the upper limit)