2021 Day 18
This was hard. I tried a few different approaches and this one eventually worked. Uses recursion, yield from (learnt about this recently), and black magic. I may add comments so that anyone reading this is slightly less confused (and so that there is hope of understanding this in the future!)