2021 Day 19
Definitely the hardest day so far. Had to look through the megathread for hints. Took the hint that offsets of overlaps were the most common distance between points in each reference frame (tried with each of the 24 orientations). Recursive dfs to find the all of the points relative to scanner zero. I used NumPy 2d arrays for rotations (which can be applied to whole lists of numbers at once)
Some functions are unused, so I may remove them at a later stage (or I may never look at this code ever again!)