2021 Day 16 - Wrap solution
Wrap the solution (both parts, optional visualisation) in a function so that I can test it with a range of inputs and snippets. Also tidy up process_tree to use a list of functions rather than if/else branches. Also use math.prod rather than reduce for the product. The version sum is now calculated using the tree structure, rather than a global counter.