2021 Day 23
23a.py shows the initial solution for part 1. 23.py contains a somewhat general possible_moves function (that behaves the same as 23a.possible_moves for size 2 rooms, but also works for larger rooms)
Memoized recursive solution (my new favourite type of solution!) that returns the lowest cost to get from here to the target_state (currently global, could be an arg instead). If it is not possible to get to the target, returns infinity.
The magic (very dark magic indeed) is the possible_moves function. It doesn't include the move that takes an amphipod straight to its room, though that move is actually a move into the hallway followed by a move into its room (so I don't need the direct move)

Couldn't be bothered parsing input, so I hardcoded it (for both parts initially, but now part 2 builds on the part 1 input)