2024 Day 5 - Uiua golf
A lot of credit to simosbingbong for this. Uses table-couple to make a grid of all the pairs in the update list. Uses memberof to find where the pairs are. This makes the same upper triangular matrices as before, without the hassle of finding all the relevant pairs and then manually locating them in a giant grid.