2024 Day 8 - Slightly more general solution
For all known inputs, dx and dy are coprime, so no antinodes appear between the nodes for part 1 (at 1/3 and 2/3). The problem states that there are two antinodes for each pair of nodes, which implies that this case never occurs for given inputs. This assumption doesn't quite extend to part 2, since that only implies that the difference isn't a multiple of 3, not that they're coprime